+++ src/main/resources/kernelGetBlockCosineSimilarity2D.cl
@@ -27,6 +27,6 @@ kernel void kernelGetBlockCosineSimilarity2D(
 
     float test_std = local_stds[gy*w+gx];
 
-    float similarity = (ref_std*test_std) / (sqrt(ref_std*ref_std)*sqrt(test_std*test_std)+EPSILON);
+    float similarity = (ref_std*test_std) / (sqrt((float)ref_std*(float)ref_std)*sqrt((float)test_std*(float)test_std)+EPSILON);
     cosine_similarity_map[gy*w+gx] = (float)fmax(0.0f, similarity);
 }
